remove func effect shifts from multidms (#38)
Addresses #18. We are not using these shifts as we get good enough
results using the simpler and more intuitive approach of simply
taking the differences between conditions, so removing from pipeline.

@juxh23, if you agree (eg, you are also not looking at the `multidms`
shift analysis), then go ahead and merge this. If you actually
think that analysis is useful, let me know and we can reconsider.