add summary plots
This pull request adds summary plots showing in one integrated interactive
display the effects of mutations on entry in all three cells and on Mxra8
binding. The data in these plots are pre-filtered, and there is also a
related CSV produced with the filtered values. This CSV can be starting
point for other analyses.

I also updated to `dms-vep-pipeline-3` version 3.25.0, and changed the color
scale to use the same cell entry colors (redblue) as Brendan did.